api/requirements.txt
Django
djangorestframework
requests
shapely
drf-yasg